Refactoring reducer
1. slice.js를 만들어 createSlice를 만들었습니다.
2. 기존 recuder의 initialState와 reducers 등을 가져왔습니다.
3. reducer test에서 actions를 slice에서 가져와서 사용하도록 변경하였습니다.
4. reducers에서 action의 payload를 실제 dispatch할 때의 parameter형태와 동일하게 변경하였습니다.
